Make blinkColorSlow more generic Now blinks a single colorBlink, accepting color and duration

+++ src/main/java/org/usfirst/frc/team2713/robot/commands/led/LEDLoopCommand.java
@@ -24,7 +24,7 @@ public class LEDLoopCommand extends Command {
     this.setInterruptible(true);
   }
   
-  public LEDLoopCommand(LEDSubsystem ledSubsystem, List<RGBValue> colors, int loops, int duration) {
+  public LEDLoopCommand(LEDSubsystem ledSubsystem, List<RGBValue> colors, int loops, double duration) {
     this.ledSubsystem = ledSubsystem;
     this.colors = colors;
     this.loops = loops;
